break down storage module
